# openapi.yml
openapi: 3.1.2
info:
  title: "短网址"
  description: "短网址接口"
  contact:
    name: "Jetsung Chan"
    url: "https://github.com/jetsung"
    email: "i@jetsung.com"
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.3.0-preview.1
servers:
  - url: http://127.0.0.1:8080
    description: Version 1
...
paths:
...
  /go/{short_code}:
    get:
      tags:
        - shorten
      summary: "短码跳转"
      description: "根据短码 308 重定向到原始 URL（无需认证），并记录一条访问历史"
      operationId: "redirectToUrl"
      parameters:
        - name: short_code
          in: path
          description: "短码"
          required: true
          schema:
            type: string
            maxLength: 16
            pattern: "^[a-zA-Z0-9]+$"
            example: "abc123"
      responses:
        "308":
          description: "永久重定向到原始 URL"
          headers:
            Location:
              description: "原始 URL"
              schema:
                type: string
                format: uri
        "404":
          description: "短码不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/account/login:
    post:
      tags:
        - account
      description: 登录接口
      operationId: login
      requestBody:
        description: 登录系统
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginParams"
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-codegen-request-body-name: body
    x-swagger-router-controller: api
  /api/account/logout:
    post:
      description: 退出登录接口
      operationId: logout
      tags:
        - account
      responses:
        "204":
          description: Success
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    x-swagger-router-controller: api
  /api/users/current:
    get:
      tags:
        - account
      description: 获取当前的用户
      operationId: currentUser
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    x-swagger-router-controller: api
  /api/oidc/login:
    get:
      tags:
        - account
      summary: "OIDC 登录"
      description: "重定向浏览器到 IdP 授权端点（302/303 Location）。需启用 OIDC（[oidc] enabled = true）；redirect_uri 由请求 Host 头推导为 {scheme}://{host}/api/oidc/callback"
      operationId: oidcLogin
      responses:
        "303":
          description: "重定向到 IdP 授权页"
          headers:
            Location:
              description: "IdP 授权端点 URL"
              schema:
                type: string
                format: uri
        "500":
          description: "OIDC 未启用或配置错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/oidc/callback:
    get:
      tags:
        - account
      summary: "OIDC 回调"
      description: "IdP 授权后的回调端点：用 code 换取令牌并校验，成功后重定向到前端 /#/dashboard"
      operationId: oidcCallback
      parameters:
        - name: code
          in: query
          description: "IdP 授权码"
          required: true
          schema:
            type: string
        - name: state
          in: query
          description: "CSRF state（含前端回跳地址与 nonce，格式 r:<redirect>|n:<nonce>）"
          required: true
          schema:
            type: string
      responses:
        "303":
          description: "重定向到前端（默认 /#/dashboard）"
          headers:
            Location:
              description: "前端回跳地址"
              schema:
                type: string
        "401":
          description: "授权失败（用户不在允许名单等）"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "OIDC 未启用或令牌交换失败"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cache/refresh:
    post:
      tags:
        - cache
      summary: "刷新缓存"
      description: "清空缓存前缀（默认 shorten:）下的所有键，并从数据库重新加载全部短链"
      operationId: "refreshCache"
      responses:
        "200":
          description: "刷新成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CacheRefreshResponse"
        "401":
          description: "未授权"
        "500":
          description: "操作失败"
        default:
          description: "未知错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/shortens:
    post:
      tags:
        - shorten
      summary: "添加短网址"
      description: "添加一个新的短网址"
      operationId: "addShorten"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Shorten"
      responses:
        "201":
          description: "短网址创建成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortenResponse"
        "400":
          description: "请求错误"
        "409":
          description: "短网址已存在"
        "500":
          description: "操作失败"
        default:
          description: "未知错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags:
        - shorten
      summary: "获取所有短址信息"
      description: "获取所有短址信息"
      operationId: "getShortens"
      parameters:
        - name: page
          in: query
          description: "页码"
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          description: "每页条数"
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sort_by
          in: query
          description: "排序字段"
          required: false
          schema:
            type: string
            default: "created_at"
            enum:
              - id
              - short_code
              - created_at
              - updated_at
        - name: order
          in: query
          description: "排序方向"
          required: false
          schema:
            type: string
            default: "desc"
            enum:
              - asc
              - desc
        - name: status
          in: query
          description: "状态"
          required: false
          schema:
            type: integer
            default: 0
            enum:
              - 0
              - 1
              - 2
        - name: short_code
          in: query
          description: "短码搜索"
          required: false
          schema:
            type: string
            maxLength: 16
        - name: original_url
          in: query
          description: "原始URL搜索（模糊匹配）"
          required: false
          schema:
            type: string
      responses:
        "200":
          description: "操作成功"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ShortenResponse"
                  meta:
                    $ref: "#/components/schemas/PageMeta"
        "400":
          description: "请求错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/shortens/batch-delete:
    post:
      tags:
        - shorten
      summary: "批量删除短网址"
      description: "批量删除短网址"
      operationId: "deleteShortenBatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchDeleteRequest"
      responses:
        "204":
          description: "操作成功"
        "400":
          description: "请求错误"
        "404":
          description: "短网址不存在"
        "500":
          description: "操作失败"
        default:
          description: "未知错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/shortens/{short_code}:
    get:
      tags:
        - shorten
      summary: "获取短网址信息"
      description: "通过短码获取短网址信息"
      operationId: "getShorten"
      parameters:
        - name: short_code
          in: path
          description: "短码"
          required: true
          schema:
            type: string
            maxLength: 16
            pattern: "^[a-zA-Z0-9]+$"
            example: "aBc123"
      responses:
        "200":
          description: "操作成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortenResponse"
        "400":
          description: "请求错误"
        "404":
          description: "短网址不存在"
        "500":
          description: "操作失败"
        "401":
          description: "未授权"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - shorten
      summary: "更新短网址"
      description: "更新一个短网址"
      operationId: "updateShorten"
      parameters:
        - name: short_code
          in: path
          description: "短码"
          required: true
          schema:
            type: string
            maxLength: 16
            pattern: "^[a-zA-Z0-9]+$"
            example: "aBc123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShortenUpdate"
      responses:
        "200":
          description: "操作成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortenResponse"
        "400":
          description: "请求错误"
        "404":
          description: "短码不存在"
        "500":
          description: "操作失败"
        default:
          description: "未知错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - shorten
      summary: "删除短网址"
      description: "删除一个短网址"
      operationId: "deleteShorten"
      parameters:
        - name: short_code
          in: path
          description: "短码"
          required: true
          schema:
            type: string
            maxLength: 16
            pattern: "^[a-zA-Z0-9]+$"
            example: "aBc123"
      responses:
        "204":
          description: "操作成功"
        "400":
          description: "请求错误"
        "404":
          description: "短网址不存在"
        "500":
          description: "操作失败"
        "401":
          description: "未授权"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/histories:
    get:
      tags:
        - history
      summary: "获取访问历史记录"
      description: "获取所有访问日志信息"
      operationId: "getHistories"
      parameters:
        - name: page
          in: query
          description: "页码"
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          description: "每页条数"
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sort_by
          in: query
          description: "排序字段"
          required: false
          schema:
            type: string
            default: "accessed_at"
            enum:
              - id
              - accessed_at
              - created_at
        - name: order
          in: query
          description: "排序方向"
          required: false
          schema:
            type: string
            default: "desc"
            enum:
              - asc
              - desc
        - name: short_code
          in: query
          description: "短码搜索"
          required: false
          schema:
            type: string
            maxLength: 16
        - name: ip_address
          in: query
          description: "IP地址搜索"
          required: false
          schema:
            type: string
      responses:
        "200":
          description: "操作成功"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/HistoryResponse"
                  meta:
                    $ref: "#/components/schemas/PageMeta"
        "400":
          description: "请求错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/histories/batch-delete:
    post:
      tags:
        - history
      summary: "批量删除访问日志"
      description: "批量删除访问日志"
      operationId: "deleteHistories"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchDeleteRequest"
      responses:
        "204":
          description: "操作成功"
        "400":
          description: "请求错误"
        "404":
          description: "日志不存在"
        "500":
          description: "操作失败"
        default:
          description: "未知错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    PageMeta:
      type: object
      required:
        - page
        - per_page
        - count
        - total
        - total_pages
      properties:
        page:
          type: integer
          minimum: 1
          description: 当前页码
          example: 1
        per_page:
          type: integer
          minimum: 1
          maximum: 100
          description: 每页数量
          example: 10
        count:
          type: integer
          minimum: 0
          description: 当前页条目数
          example: 10
        total:
          type: integer
          minimum: 0
          description: 总条目数
          example: 100
        total_pages:
          type: integer
          minimum: 0
          description: 总页数
          example: 10
    LoginParams:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: 用户名
        password:
          type: string
          format: password
          description: 密码
        auto_login:
          type: boolean
          description: 是否自动登录
          default: false
    LoginResult:
      type: object
      properties:
        token:
          type: string
          description: 登录成功后返回的 token
        error_code:
          type: string
          description: 错误码
        error_message:
          type: string
          description: 错误信息
    CurrentUser:
      type: object
      properties:
        name:
          type: string
          description: 用户名
    ErrorResponse:
      required:
        - error_code
      type: object
      properties:
        error_code:
          type: string
          description: 业务约定的错误码
        error_message:
          type: string
          description: 错误信息
    Shorten:
      type: object
      required:
        - original_url
        - short_code
      properties:
        original_url:
          type: string
          format: uri
          description: "原始长网址"
          example: "https://example.com/very/long/url"
        short_code:
          type: string
          description: "自定义短码"
          maxLength: 16
          minLength: 3
          pattern: "^[a-zA-Z0-9]+$"
          example: "abc123"
        description:
          type: string
          description: "短链描述"
          maxLength: 255
    ShortenUpdate:
      type: object
      required:
        - original_url
      properties:
        original_url:
          type: string
          format: uri
          description: "原始长网址"
        description:
          type: string
          description: "短链描述"
    ShortenResponse:
      type: object
      properties:
        id:
          type: integer
          description: "数据库 ID"
        short_code:
          type: string
          description: "短码"
          maxLength: 16
          pattern: "^[a-zA-Z0-9]+$"
          example: "abc123"
        short_url:
          type: string
          format: uri
          description: "完整短网址。short_url 未设置时形如 http://host:port/go/abc123（推断地址 + /go 前缀）；显式配置专用短域名时形如 https://s.example.com/abc123（无前缀）"
          example: "http://localhost:8080/go/abc123"
        original_url:
          type: string
          format: uri
          description: "原始长网址"
        description:
          type: string
          description: "短链描述"
        status:
          type: integer
          description: "状态：0=启用, 1=禁用, 2=未知"
          enum: [0, 1, 2]
          default: 0
        created_at:
          type: string
          format: date-time
          description: "创建时间 (ISO 8601 格式)"
          example: "2024-01-15T08:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: "更新时间 (ISO 8601 格式)"
          example: "2024-01-15T08:30:00Z"
    BatchDeleteRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          items:
            type: integer
            format: int64
          description: "要删除的 ID 列表"
          example: [1, 2, 3, 4, 5]
    CacheRefreshResponse:
      type: object
      properties:
        cleared_keys:
          type: integer
          format: int64
          description: "清除的旧缓存键数量"
          example: 3
        warmed_urls:
          type: integer
          format: int64
          description: "重新缓存的短链数量"
          example: 3
...